Make ArmaCallContext private
