+++ arma-rs/src/testing.rs
@@ -2,7 +2,7 @@
 
 use std::time::Duration;
 
-use crate::{context, Context, State, Value};
+use crate::{ArmaCallContext, Caller, Context, Mission, Server, Source, State, Value};
 
 /// Wrapper around [`crate::Extension`] used for testing.
 pub struct Extension(crate::Extension);
@@ -71,9 +71,12 @@ impl Extension {
         &self,
         function: &str,
         args: Option<Vec<String>>,
-        context: context::ArmaCallContext,
+        caller: Caller,
+        source: Source,
+        mission: Mission,
+        server: Server,
     ) -> (String, libc::c_int) {
-        self.set_arma_call_context(context);
+        self.set_arma_call_context(ArmaCallContext::new(caller, source, mission, server));
         self.handle_call(function, args)
     }
 
@@ -83,11 +86,11 @@ impl Extension {
     /// # Safety
     /// This function is unsafe because it interacts with the C API.
     pub unsafe fn call(&self, function: &str, args: Option<Vec<String>>) -> (String, libc::c_int) {
-        self.set_arma_call_context(context::ArmaCallContext::default());
+        self.set_arma_call_context(ArmaCallContext::default());
         self.handle_call(function, args)
     }
 
-    fn set_arma_call_context(&self, ctx: context::ArmaCallContext) {
+    fn set_arma_call_context(&self, ctx: ArmaCallContext) {
         self.0.arma_call_ctx.replace(ctx);
     }
